CH2080 One connection, multiple documents
